add working prototype
This change adds a working prototype for an agent and controller to
modify containerd's cri registry configuration.

The prototype consists of two components an agent and controller. The
agent is responsible of setting _registry.config_path_, restarting
containerd and copying registry files under the new _config_path_. The
controller watches if the agent is up and running on a node and removes
a taint which allows other pods to be scheduled after containerd is
restarted.

Signed-off-by: xinau <felix@ehrenpfort.de>